requirements.txt
pycorenlp
requests
networkx
spacy
pyspotlight
neo4j
python_cypher
sparqlwrapper